+++ src/radical/utils/__init__.py
@@ -1,6 +1,6 @@
 
 # allow star import, unused symbols
-# flake8: noqa
+# flake8: noqa: F401
 
 __author__    = "Radical.Utils Development Team"
 __copyright__ = "Copyright 2013, RADICAL@Rutgers"
@@ -26,7 +26,7 @@ from .futures        import NEW, RUNNING, DONE, FAILED, CANCELED
 from .url            import Url
 from .host           import get_hostip, get_hostname
 from .host           import get_hostlist, get_hostlist_by_range
-from .host           import create_hostfile, compress_hostlist
+from .host           import create_hostfile, compress_hostlist, is_localhost
 from .dict_mixin     import DictMixin, dict_merge, dict_stringexpand, dict_diff
 from .dict_mixin     import PRESERVE, OVERWRITE, iter_diff
 from .lockable       import Lockable
@@ -44,7 +44,8 @@ from .shell          import sh_callout, sh_callout_bg, sh_callout_async
 from .testing        import sys_exit
 from .testing        import TestConfig
 from .testing        import set_test_config, add_test_config, get_test_config
-from .env            import env_prep, env_eval, env_read, env_read_lines, env_diff
+from .env            import env_read, env_write, env_read_lines, env_eval
+from .env            import env_prep, env_diff, EnvProcess
 from .stack          import stack
 
 from .zmq            import Bridge