+++ src/radical/utils/daemon.py
@@ -1,3 +1,4 @@
+# pylint: disable=try-except-raise
 
 import io
 import os
@@ -6,12 +7,11 @@ import queue
 import signal
 import multiprocessing as mp
 
+from .misc    import ru_open
 from .testing import sys_exit
 
 # from
 # http://stackoverflow.com/questions/1417631/python-code-to-daemonize-a-process
-
-# pylint: disable=try-except-raise
 
 
 # ------------------------------------------------------------------------------
@@ -65,9 +65,9 @@ def daemonize(main=None, args=None, stdout=None, stderr=None, stdin=None,
         raise RuntimeError(
             'Failed to start daemon: %d (%s)\n' % (e.errno, e.strerror)) from e
 
-
-    # decouple from parent process group
-    os.setsid()
+    # decoupling from parent process group is disabled
+    # (some launch methods, e.g., APRun, required to run within the same group)
+    #    os.setsid()
 
     # second fork
     try:
@@ -87,26 +87,26 @@ def daemonize(main=None, args=None, stdout=None, stderr=None, stdin=None,
     # redirect standard file descriptors
     if stdin:
         try:
-            si = open(stdin, 'r')
+            si = ru_open(stdin, 'r')
             os.dup2(si.fileno(), sys.stdin.fileno())
         except io.UnsupportedOperation:
-            sys.stdin = open(stdin, 'r')
+            sys.stdin = ru_open(stdin, 'r')
 
     if stdout:
         try:
             sys.stdout.flush()
-            so = open(stdout, 'a+')
+            so = ru_open(stdout, 'a+')
             os.dup2(so.fileno(), sys.stdout.fileno())
         except io.UnsupportedOperation:
-            sys.stdout = open(stdout, 'a+')
+            sys.stdout = ru_open(stdout, 'a+')
 
     if stderr:
         try:
             sys.stderr.flush()
-            se = open(stderr, 'a+')
+            se = ru_open(stderr, 'a+')
             os.dup2(se.fileno(), sys.stderr.fileno())
         except io.UnsupportedOperation:
-            sys.stderr = open(stderr, 'a+')
+            sys.stderr = ru_open(stderr, 'a+')
 
     if main:
         # we are successfully daemonized - run the workload and exit
@@ -115,7 +115,7 @@ def daemonize(main=None, args=None, stdout=None, stderr=None, stdin=None,
         sys_exit(0)
 
     else:
-        # just return - the callinng code will now continue daemonized
+        # just return - the calling code will now continue daemonized
         return