+++ src/radical/utils/host.py
@@ -10,7 +10,7 @@ import socket
 
 from functools import reduce
 
-from .misc import as_list
+from .misc import as_list, ru_open
 
 
 # ------------------------------------------------------------------------------
@@ -24,12 +24,7 @@ def get_hostname():
     global _hostname                                     # pylint: disable=W0603
     if not _hostname:
 
-        _hostname = socket.gethostname()
-        if '.' in _hostname:
-            try:
-                _hostname = socket.gethostbyaddr(_hostname)[0]
-            except socket.herror:
-                pass
+        _hostname = socket.getfqdn()
 
     return _hostname
 
@@ -119,7 +114,7 @@ def create_hostfile(sandbox, name, hostlist, sep=' ', impaired=False):
 
     hostlist = as_list(hostlist)
     filename = '%s/%s.hosts' % (sandbox or '.', name)
-    with open(filename, 'w') as fout:
+    with ru_open(filename, 'w', encoding='utf8') as fout:
 
         if not impaired:
             # create dict: {'host1': x, 'host2': y}
@@ -256,4 +251,28 @@ def get_hostlist(hoststring):
     return output
 
 
+# --------------------------------------------------------------------------
+#
+def is_localhost(host: str) -> bool:
+    '''
+    Returns `True` if given hostname is localhost, `False` otherwise.
+    '''
+
+    if not host:
+        return True
+
+    elif host == 'localhost':
+        return True
+
+    else:
+        sockhost = socket.gethostname()
+        while sockhost:
+            if host == sockhost:
+                return True
+            sockhost = '.'.join(sockhost.split('.')[1:])
+
+    return False
+
+
 # ------------------------------------------------------------------------------
+