+++ tests/unittests/test_host.py
@@ -5,6 +5,8 @@ __copyright__ = 'Copyright 2021, The RADICAL-Cybertools Team'
 __license__   = 'MIT'
 
 import os
+import socket
+
 os.environ['RADICAL_BASE'] = '/tmp'
 
 import glob
@@ -45,17 +47,26 @@ class HostTestCase(TestCase):
     #
     @mock.patch('radical.utils.host.socket')
     def test_get_hostname(self, mocked_socket):
+        """Check caching wrapper for gethostname."""
 
-        mocked_socket.gethostname.return_value   = '127.0.0.1'
-        mocked_socket.gethostbyaddr.return_value = ('localhost', None, None)
+        # Note that ru.get_hostname() actually uses getfqdn()
+        fq_hostname = socket.getfqdn()
+        mocked_socket.gethostname.return_value = socket.gethostname()
+        mocked_socket.getfqdn.return_value = fq_hostname
 
+        ru.host._hostname = None                         # pylint: disable=W0212
         self.assertFalse(mocked_socket.gethostname.called)
-        self.assertEqual(ru.get_hostname(), 'localhost')
-        self.assertTrue(mocked_socket.gethostname.called)
+        self.assertFalse(mocked_socket.getfqdn.called)
+        self.assertEqual(ru.get_hostname(), fq_hostname)
+        self.assertTrue(mocked_socket.getfqdn.called)
 
-        mocked_socket.gethostname.reset_mock()
-        ru.get_hostname()  # `socket.gethostname` is not called 2nd time
+        mocked_socket.getfqdn.reset_mock()
+        ru.get_hostname()  # System level hostname lookup is not called 2nd time
         self.assertFalse(mocked_socket.gethostname.called)
+        self.assertFalse(mocked_socket.getfqdn.called)
+
+        # reset
+        ru.host._hostname = None                         # pylint: disable=W0212
 
     # --------------------------------------------------------------------------
     #
@@ -70,7 +81,7 @@ class HostTestCase(TestCase):
 
             filename = ru.create_hostfile(self._base_dir, 'tc', hostlist, sep)
 
-            with open(filename) as f:
+            with open(filename, 'r', encoding='utf8') as f:
                 hostfile_lines = f.readlines()
             self.assertEqual(test_case['result']['lines'], hostfile_lines)